% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_price.R
\name{total_price}
\alias{total_price}
\title{Calculates the price of}
\usage{
total_price(order, currency = "EUR")
}
\arguments{
\item{order}{is data frame of electronic products.
Products have prices in different currencies.}

\item{currency}{Currency of interest.}
}
\value{
Sum of all products’ prices from the given order,
after applying a discount.
}
\description{
Calculates the price of
}
